// src/api/consult/consult.types.d.ts
export type TalkLaterAdviserSpecialtyData = TalkLaterAdviserPivot & {
    id: number;
    created_at: string;
    updated_at: string;
    speciality: TalkLaterAdviserSpecialty;
};
// ...
export declare enum CallMethod {
    Audio = "Audio",
    Video = "Video"
}
// ...
export type DeviceInfo = {
    deviceType: string;
    osVersion: string;
};
export type CardDetails = {
    id: number;
    brand: string | null;
    card_no: string;
    country: string | null;
    is_used: 0 | 1;
    lead_id: string;
    card_token: string;
    created_at: string;
    updated_at: string;
    expiry_year: string;
    expiry_month: string;
};
export type ConsultUserDetails = {
    id: string;
    createdAt: string;
    updatedAt: string;
    phoneNumber: string;
    name: string;
    profilePicture: string | null;
    gender: string;
    dob: string | null;
    userType: string;
    doctor_prefix: string;
    doctorType: string | null;
    doctorQualification: string;
    quality: string;
    doctorSpecialties: TalkLaterAdviserSpecialtyData[];
    doctorRegistrationId: string | null;
    registration_id: string;
    language: string;
    doctorSignatureUrl: string;
    email: string;
    onlineStatus: 'OFFLINE';
    socket: string | null;
    availabilityStatus: any;
    roleId: number;
    userId: number;
    token: string;
    pushNotificationToken: string | null;
    refreshToken: string;
    healthData: null;
    totalConsultation: number;
    averageRating: number;
    totalRating: number;
    totalRatingCount: number;
    isHealthDataPermissionAccepted: boolean;
    doctor_preference: string | null;
    device_os: string | null;
    lead_id: number | null;
    doc_speciality_id: number | null;
    doc_speciality_title: string | null;
    doc_speciality_icon: string | null;
    doc_time_slot: string | null;
    doc_consultation_fee: string | null;
    lead_pkg_id: number | null;
    isUserSync: boolean;
};
export declare enum CallEndedType {
    Dropped = "Dropped",
    Consulted = "Consulted",
    InCompleteConsult = "In Complete Consult",
    Cancelled = "Cancelled",
    Initiated = "Initiated",
    Scheduled = "Scheduled",
    Missed = "Missed",
    Rescheduled = "Rescheduled",
    Offline = "Offline"
}
export declare enum CallEndedTypeId {
    Dropped = 1,
    Consulted = 2,
    InCompleteConsult = 3,
    Cancelled = 4,
    Initiated = 5,
    Scheduled = 6,
    Missed = 7,
    Rescheduled = 8,
    Offline = 9
}
export type ConsultationDetails = {
    id: string;
    createdAt: string;
    updatedAt: string;
    consultationId: number;
    flashConsultationId: number | null;
    flashConsultationChargeId: string;
    repId: null;
    flashAppointmentId: number;
    isConsultationCharged: boolean;
    consultationStatus: string;
    callEndedType: CallEndedType;
    callEndedTypeId: CallEndedTypeId;
    consultationType: string | null;
    consultationModeId: 2;
    consultationMode: string;
    consultationSubMode: string;
    organizationId: number | null;
    organizationName: string | null;
    packageId: number;
    price: number;
    discount: number;
    flashCredit: number;
    is_flash_point: number;
    medOrderId: number | null;
    labOrderId: number | null;
    gender: string;
    age: string;
    callType: CallMethod;
    consultationFor: string;
    consultationForName: string;
    consultationForAge: string;
    consultationForGender: string;
    isHealthDataRequested: boolean;
    isMedOrderPlaced: boolean;
    isLabOrderPlaced: boolean;
    unlistedMedications: null;
    medications: Medication[] | null;
    labTests: LabTestDetails[] | null;
    diagnoses: ClinicalCommonDetails | null;
    symptoms: ClinicalCommonDetails | null;
    allergies: ClinicalCommonDetails | null;
    medPreFiles: string[] | null;
    prescription: string | null;
    deviceInfo: DeviceInfo | null;
    cardDetails: CardDetails;
    addressId: number;
    cardId: number;
    requestQuote: null;
    requestedAt: string | null;
    startedAt: string | null;
    endedAt: string | null;
    appointmentDate: string;
    appointmentStatus: string;
    appointmentType: string | null;
    totalConsultationTime: string | null;
    totalConsultationTimeInSecond: string | null;
    language: string;
    duration: string;
    is_join: boolean;
    isRequestedForCall: boolean;
    remainingSeconds: number;
    exceededSeconds: number;
    isExceeded: boolean;
    isConsultationOver: boolean;
    is_refund_requested: boolean;
    lead_pkg_id: number | null;
    merMustChecks: null;
    merPackageId: number | null;
    totalClaimCredits: number;
    isClaim: boolean;
    cancelReason: string | null;
    hasFailedPrescription: boolean;
    doctor: ConsultUserDetails | null;
    patient: ConsultUserDetails;
    medicalAssessment: null;
    appointmentDateString: string;
};
export type ClinicalCommonDetails = {
    items: ClinicalCommonDataDetails[];
    note: string;
};
export type LabTestDetails = {
    data: any;
    note: string;
};
export type Medication = {
    dosage: string;
    duration: string;
    frequency: string;
    medicineId: string;
    medicineName: string;
    notes: string;
    route: string;
    schedules: Schedule[];
    timing: string;
};
export type Schedule = {
    count: number;
    id: string;
    status: boolean;
    title: string;
};
// ...
export type ClinicalCommonDataDetails = {
    id: string;
    name: string;
    createdAt: string;
    updatedAt: string;
};